pyproject.toml
[tool.poetry]
name = "rag-test"
version = "0.1.0"
description = ""
authors = ["Khant Htoo Naing <khanthtoonaing2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
langchain-huggingface = "^0.0.3"
langchain-community = "^0.2.4"
python-dotenv = "^1.0.1"
langchain = "^0.2.3"
faiss-cpu = "^1.8.0"
langchainhub = "^0.1.18"
streamlit = "^1.35.0"
wikipedia = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
